+++ frontend/src/main.jsx
@@ -1,16 +1,16 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { UserProvider } from "./context/UserContext.jsx";
-import { ScrapeConfigProvider } from "./context/RetrievalInstructionsContext.jsx";
+import { RetrievalInstructionProvider } from "./context/RetrievalInstructionContext.jsx";
 import App from './App.jsx';
 import '../index.css';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <UserProvider>
-      <ScrapeConfigProvider>
+      <RetrievalInstructionProvider>
         <App/>
-      </ScrapeConfigProvider>
+      </RetrievalInstructionProvider>
     </UserProvider>
   </StrictMode>
 )
